Modularized code to be used for webpack
